+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * internal: improve Gradle build performance.
 * internal: migrate Gradle build to Kotlin DSL.
 * add an option to adjust UI Scale Factor for user icons. Useful if you run IDE with `-Dsun.java2d.uiScale.enabled=false` flag.
+* support `KNOWN_ISSUES` and `OPEN_ISSUES` files.
 
 ## 1.38.0 (2020/09/30)
 * rework `package-info.java` files.