+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * internal: migrate Gradle build to Kotlin DSL.
 * add an option to adjust UI Scale Factor for user icons. Useful if you run IDE with `-Dsun.java2d.uiScale.enabled=false` flag.
 * support `KNOWN_ISSUES` and `OPEN_ISSUES` files.
+* settings: minor UI improvements.
 
 ## 1.38.0 (2020/09/30)
 * rework `package-info.java` files.