+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Extra Icons Change Log
 
-## 1.39.0 (WIP)
+## 1.39.0 (2020/10/30)
 * support [Mockk](https://mockk.io/#settings-file) `io/mockk/settings.properties` files. 
 * support [gradle-intellij-plugin](https://github.com/JetBrains/gradle-intellij-plugin) `idea-sandbox` folders.
 * internal: improve Gradle build performance.